refactor(server): colocate lib and service tests
