+++ packages/emitsignal-server/src/lib/email-service.spec.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it, mock } from 'bun:test';
 
-import { EmailService } from '#/lib/email-service';
+import { EmailService } from './email-service';
 
 describe('EmailService', () => {
     it('init stores the queue for later use', async () => {
