+++ packages/emitsignal-server/src/services/topic-access.spec.ts
@@ -4,7 +4,7 @@ import { prismaMock } from '#/__tests__/mocks';
 
 mock.module('#/lib/prisma', () => ({ prisma: prismaMock }));
 
-import { resolveTopicCapabilities } from '#/services/topic-access';
+import { resolveTopicCapabilities } from './topic-access';
 
 describe('resolveTopicCapabilities', () => {
     beforeEach(() => {
