+++ packages/emitsignal-server/src/utils/media-refs.spec.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'bun:test';
 
-import { normalizeMediaInput, validateMessageMedia } from '#/utils/media-refs';
+import { normalizeMediaInput, validateMessageMedia } from './media-refs';
 
 describe('normalizeMediaInput', () => {
     it('returns empty for null/undefined', () => {
